requirements.txt
bilibili-api-python>=16.0.0
aiohttp>=3.8.0
aiofiles>=23.0.0
